feat(TDI-40715): Rewrite unit tests for services
